+++ tests/test_solver.py
@@ -37,7 +37,9 @@ class TestSolve:
         sfield = fields.get_source_field(**dat['input_source'])
 
         # F-cycle
-        efield = solver.solve(model, sfield, verb=4)
+        efield = solver.solve(model, sfield, sslsolver=False,
+                              semicoarsening=False, linerelaxation=False,
+                              verb=4)
         out, _ = capsys.readouterr()
 
         assert ' emg3d START ::' in out
@@ -55,20 +57,25 @@ class TestSolve:
         assert_allclose(dat['Fresult'], efield)
 
         # W-cycle
-        wfield = solver.solve(model, sfield, cycle='W')
+        wfield = solver.solve(model, sfield, sslsolver=False,
+                              semicoarsening=False, linerelaxation=False,
+                              cycle='W')
 
         # Check all fields (ex, ey, and ez)
         assert_allclose(dat['Wresult'], wfield)
 
         # V-cycle
-        vfield = solver.solve(model, sfield, cycle='V')
+        vfield = solver.solve(model, sfield, sslsolver=False,
+                              semicoarsening=False, linerelaxation=False,
+                              cycle='V')
         _, _ = capsys.readouterr()  # clear output
 
         # Check all fields (ex, ey, and ez)
         assert_allclose(dat['Vresult'], vfield)
 
         # BiCGSTAB with some print checking.
-        efield = solver.solve(model, sfield, verb=4, sslsolver=True)
+        efield = solver.solve(model, sfield, verb=4, semicoarsening=False,
+                              linerelaxation=False)
         out, _ = capsys.readouterr()
         assert ' emg3d START ::' in out
         assert ' [hh:mm:ss] ' in out
@@ -82,7 +89,9 @@ class TestSolve:
         assert_allclose(dat['bicresult'], efield)
 
         # Same as previous, without BiCGSTAB, but some print checking.
-        efield = solver.solve(model, sfield, verb=4)
+        efield = solver.solve(model, sfield, sslsolver=False,
+                              semicoarsening=False, linerelaxation=False,
+                              verb=4)
         out, _ = capsys.readouterr()
         assert ' emg3d START ::' in out
         assert ' [hh:mm:ss] ' in out
@@ -94,7 +103,8 @@ class TestSolve:
         # Max it
         maxit = 2
         _, info = solver.solve(
-                model, sfield, verb=3, maxit=maxit, return_info=True)
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, verb=3, maxit=maxit, return_info=True)
         out, _ = capsys.readouterr()
         assert ' MAX. ITERATION REACHED' in out
         assert maxit == info['it_mg']
@@ -102,17 +112,22 @@ class TestSolve:
         assert 'MAX. ITERATION REACHED' in info['exit_message']
 
         # BiCGSTAB with lower verbosity, print checking.
-        _ = solver.solve(model, sfield, verb=3, maxit=1, sslsolver=True)
+        _ = solver.solve(model, sfield, semicoarsening=False,
+                         linerelaxation=False, verb=3, maxit=1)
         out, _ = capsys.readouterr()
         assert ' MAX. ITERATION REACHED' in out
 
         # Just check if it runs without failing for other solvers.
-        _ = solver.solve(model, sfield, verb=5, maxit=1, sslsolver='gcrotmk')
+        _ = solver.solve(model, sfield, sslsolver='gcrotmk',
+                         semicoarsening=False, linerelaxation=False, verb=5,
+                         maxit=1)
 
         # Provide initial field.
         _, _ = capsys.readouterr()  # empty
         efield_copy = efield.copy()
-        outarray = solver.solve(model, sfield, efield_copy, verb=3)
+        outarray = solver.solve(
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, efield=efield_copy, verb=3)
         out, _ = capsys.readouterr()
 
         # Ensure there is no output.
@@ -122,7 +137,9 @@ class TestSolve:
         assert_allclose(efield, efield_copy)
 
         # Provide initial field and return info.
-        info = solver.solve(model, sfield, efield_copy, return_info=True)
+        info = solver.solve(
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, efield=efield_copy, return_info=True)
         assert info['it_mg'] == 0
         assert info['it_ssl'] == 0
         assert info['exit'] == 0
@@ -132,9 +149,7 @@ class TestSolve:
         # without linerelaxation nor semicoarsening.
         _, _ = capsys.readouterr()  # empty
         efield = fields.Field(grid)
-        outarray = solver.solve(
-                model, sfield, efield, sslsolver=True, semicoarsening=True,
-                linerelaxation=True, maxit=2, verb=4)
+        outarray = solver.solve(model, sfield, efield=efield, maxit=2, verb=4)
         out, _ = capsys.readouterr()
         assert "after                       1 F-cycles    4 1" in out
         assert "after                       2 F-cycles    5 2" in out
@@ -143,15 +158,20 @@ class TestSolve:
         wrong_sfield = fields.Field(grid)
         wrong_sfield.field = sfield.field
         with pytest.raises(ValueError, match="Source field is missing frequ"):
-            solver.solve(model, wrong_sfield, efield=efield, verb=1)
+            solver.solve(model, wrong_sfield, sslsolver=False,
+                         semicoarsening=False, linerelaxation=False,
+                         efield=efield, verb=1)
 
         # Check stagnation by providing an almost zero source field.
-        _ = solver.solve(model, sfield*0+1e-20, maxit=100)
+        _ = solver.solve(model, sfield*0+1e-20, sslsolver=False,
+                         semicoarsening=False, linerelaxation=False, maxit=100)
         out, _ = capsys.readouterr()
         assert "STAGNATED" in out
 
         # Check a zero field is returned for a zero source field.
-        efield = solver.solve(model, sfield*0, maxit=100, verb=3)
+        efield = solver.solve(
+                model, sfield*0, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, maxit=100, verb=3)
         out, _ = capsys.readouterr()
         assert "RETURN ZERO E-FIELD (provided sfield is zero)" in out
         assert np.linalg.norm(efield) == 0.0
@@ -166,18 +186,24 @@ class TestSolve:
             inp[n] = int(inp[n])
         inp['verb'] = 5
 
-        efield = solver.solve(model, sfield, **inp)
+        efield = solver.solve(model, sfield, sslsolver=False, **inp)
 
         assert_allclose(dat['result'], efield.field)
 
         _, _ = capsys.readouterr()  # Clean up
 
         # Check with provided e-field; 2x2 iter should yield the same as 4 iter
-        efield2 = solver.solve(model, sfield, maxit=4, verb=0)
+        efield2 = solver.solve(
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, maxit=4, verb=0)
         out, _ = capsys.readouterr()  # Clean up
         assert "* WARNING :: MAX. ITERATION REACHED, NOT CONVERGED" in out
-        efield3 = solver.solve(model, sfield, maxit=2, verb=0)
-        solver.solve(model, sfield, efield3, maxit=2, verb=0)
+        efield3 = solver.solve(
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, maxit=2, verb=0)
+        solver.solve(
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False, efield=efield3, maxit=2, verb=0)
 
         assert_allclose(efield2, efield3)
 
@@ -185,11 +211,9 @@ class TestSolve:
 
         # One test without post-smoothing to check if it runs.
         efield4 = solver.solve(
-                model, sfield, sslsolver=True, semicoarsening=True,
-                linerelaxation=True, maxit=20, nu_pre=0, nu_post=4, verb=4)
+                model, sfield, maxit=20, nu_pre=0, nu_post=4, verb=4)
         efield5 = solver.solve(
-                model, sfield, sslsolver=True, semicoarsening=True,
-                linerelaxation=True, maxit=20, nu_pre=4, nu_post=0, verb=4)
+                model, sfield, maxit=20, nu_pre=4, nu_post=0, verb=4)
         # They don't converge, and hence don't agree. Just a lazy test.
         assert_allclose(efield4, efield5, atol=1e-15, rtol=1e-5)
 
@@ -201,7 +225,9 @@ class TestSolve:
                 origin=np.array([-0.5, -1, -1]))
         sfield = alternatives.get_source_field(mesh, [0, 0, 0, 0, 0], 1)
         model = models.Model(mesh)
-        _ = solver.solve(model, sfield, verb=4, nu_pre=0)
+        _ = solver.solve(
+                model, sfield, sslsolver=False, linerelaxation=False,
+                semicoarsening=False, verb=4, nu_pre=0)
         out, _ = capsys.readouterr()
         assert "(Cycle-QC restricted to first 70 steps of 72 steps.)" in out
         assert "DIVERGED" in out
@@ -237,13 +263,16 @@ class TestSolve:
         sfield = fields.get_source_field(**dat['input_source'])
 
         # F-cycle
-        efield = solver.solve(model, sfield)
+        efield = solver.solve(
+                model, sfield, sslsolver=False, semicoarsening=False,
+                linerelaxation=False)
 
         # Check all fields (ex, ey, and ez)
         assert_allclose(dat['Fresult'], efield, atol=1e-14)
 
         # BiCGSTAB with some print checking.
-        efield = solver.solve(model, sfield, sslsolver=True)
+        efield = solver.solve(
+                model, sfield, semicoarsening=False, linerelaxation=False)
 
         # Check all fields (ex, ey, and ez)
         assert_allclose(dat['bicresult'], efield, atol=1e-14)
@@ -252,41 +281,124 @@ class TestSolve:
         efield = fields.Field(grid, dtype=np.complex_)
 
         with pytest.raises(ValueError, match='Source field and electric fiel'):
-            efield = solver.solve(model, sfield, efield=efield)
+            efield = solver.solve(
+                    model, sfield, sslsolver=False, semicoarsening=False,
+                    linerelaxation=False, efield=efield)
 
 
-def multigrid():
-    pass
-    # No test at the moment. Add one!
+class TestMultigrid:
+    # Everything should be tested just fine in `test_solver`. Just check here
+    # that all code is reached.
+
+    def test_basic(self, capsys):
+        # This should reach every line of solver.multigrid.
+        dat = REGRES['res']
+        model = models.Model(**dat['input_model'])
+        grid = model.grid
+        sfield = fields.get_source_field(**dat['input_source'])
+        vmodel = models.VolumeModel(model, sfield)
+        efield = fields.Field(grid)  # Initiate e-field.
+
+        # Get var-instance
+        var = solver.MGParameters(
+                cycle='F', sslsolver=False, semicoarsening=True,
+                linerelaxation=True, shape_cells=grid.shape_cells, verb=5,
+                nu_init=2, maxit=-1,
+        )
+        var.l2_refe = sl.norm(sfield, check_finite=False)
+
+        # Call multigrid.
+        solver.multigrid(vmodel, sfield, efield, var)
+        out, _ = capsys.readouterr()
+        assert '> CONVERGED' in out
 
 
-def test_krylov(capsys):
+class TestKrylov:
+    # Everything should be tested just fine in `test_solver`. Just check here
+    # for bicgstab-error, and that all code is reached.
 
-    # Everything should be tested just fine in `test_solver`.
-    # Just check here for bicgstab-error.
+    def test_bicgstab_error(self, capsys):
+        # Load any case.
+        dat = REGRES['res']
+        model = models.Model(**dat['input_model'])
+        grid = model.grid
+        model.property_x /= 100000  # Set stupid input to make bicgstab fail.
+        model.property_y *= 100000  # Set stupid input to make bicgstab fail.
+        sfield = fields.get_source_field(**dat['input_source'])
+        vmodel = models.VolumeModel(model, sfield)
+        efield = fields.Field(grid)  # Initiate e-field.
 
-    # Load any case.
-    dat = REGRES['res']
-    model = models.Model(**dat['input_model'])
-    grid = model.grid
-    model.property_x /= 100000  # Set stupid input to make bicgstab fail.
-    model.property_y *= 100000  # Set stupid input to make bicgstab fail.
-    sfield = fields.get_source_field(**dat['input_source'])
-    vmodel = models.VolumeModel(model, sfield)
-    efield = fields.Field(grid)  # Initiate e-field.
+        # Get var-instance
+        var = solver.MGParameters(
+                cycle=None, sslsolver=True, semicoarsening=False,
+                linerelaxation=False, shape_cells=grid.shape_cells, verb=3,
+                maxit=-1,
+        )
+        var.l2_refe = sl.norm(sfield, check_finite=False)
 
-    # Get var-instance
-    var = solver.MGParameters(
-            cycle=None, sslsolver=True, semicoarsening=False,
-            linerelaxation=False, shape_cells=grid.shape_cells, verb=4,
-            maxit=-1,
-    )
-    var.l2_refe = sl.norm(sfield, check_finite=False)
+        # Call krylov and ensure it fails properly.
+        solver.krylov(vmodel, sfield, efield, var)
+        out, _ = capsys.readouterr()
+        assert '* ERROR   :: Error in bicgstab' in out
 
-    # Call krylov and ensure it fails properly.
-    solver.krylov(vmodel, sfield, efield, var)
-    out, _ = capsys.readouterr()
-    assert '* ERROR   :: Error in bicgstab' in out
+    def test_cycle_gcrotmk(self, capsys):
+
+        # Load any case.
+        dat = REGRES['res']
+        model = models.Model(**dat['input_model'])
+        grid = model.grid
+        sfield = fields.get_source_field(**dat['input_source'])
+        vmodel = models.VolumeModel(model, sfield)
+        efield = fields.Field(grid)  # Initiate e-field.
+
+        # Get var-instance
+        var = solver.MGParameters(
+                cycle='F', sslsolver='gcrotmk', semicoarsening=False,
+                linerelaxation=False, shape_cells=grid.shape_cells, verb=4,
+                maxit=5,
+        )
+        var.l2_refe = sl.norm(sfield, check_finite=False)
+
+        # Call krylov and ensure it fails properly.
+        solver.krylov(vmodel, sfield, efield, var)
+        out, _ = capsys.readouterr()
+        assert 'DIVERGED' in out
+
+    def test_cycle(self, capsys):
+
+        # Load any case.
+        dat = REGRES['res']
+        model = models.Model(**dat['input_model'])
+        grid = model.grid
+        sfield = fields.get_source_field(**dat['input_source'])
+        vmodel = models.VolumeModel(model, sfield)
+        efield = fields.Field(grid)  # Initiate e-field.
+
+        # Get var-instance
+        var = solver.MGParameters(
+                cycle='F', sslsolver=True, semicoarsening=True,
+                tol=1.0, linerelaxation=True, shape_cells=grid.shape_cells,
+                verb=4, maxit=2,
+        )
+        var.l2_refe = sl.norm(sfield, check_finite=False)
+
+        # Call krylov and ensure it fails properly.
+        solver.krylov(vmodel, sfield, efield, var)
+        out, _ = capsys.readouterr()
+        assert '> CONVERGED' in out
+
+        # Call krylov and ensure it fails properly.
+        efield = fields.Field(grid)  # Initiate e-field.
+        # Get var-instance
+        var = solver.MGParameters(
+                cycle='F', sslsolver=True, semicoarsening=False,
+                linerelaxation=False, shape_cells=grid.shape_cells, verb=4,
+                maxit=1,
+        )
+        var.l2_refe = sl.norm(sfield, check_finite=False)
+        solver.krylov(vmodel, sfield, efield, var)
+        out, _ = capsys.readouterr()
+        assert 'MAX. ITERATION REACHED' in out
 
 
 def test_smoothing():
@@ -1014,20 +1126,23 @@ def test_print_one_liner(capsys):
 
     # Dynamic one-liner.
     out, _ = capsys.readouterr()
-    _ = solver.solve(model, sfield, verb=1)
+    _ = solver.solve(model, sfield, sslsolver=False, semicoarsening=False,
+                     linerelaxation=False, verb=1)
     out, _ = capsys.readouterr()
     assert '6; 0:00:' in out
     assert '; CONVERGED' in out
 
     out, _ = capsys.readouterr()
-    _ = solver.solve(model, sfield, sslsolver=True, verb=1)
+    _ = solver.solve(model, sfield, sslsolver=True, semicoarsening=False,
+                     linerelaxation=False, verb=1)
     out, _ = capsys.readouterr()
     assert '3(5); 0:00:' in out
     assert '; CONVERGED' in out
 
     # One-liner.
     out, _ = capsys.readouterr()
-    _ = solver.solve(model, sfield, sslsolver=True, verb=1)
+    _ = solver.solve(model, sfield, sslsolver=True, semicoarsening=False,
+                     linerelaxation=False, verb=1)
     out, _ = capsys.readouterr()
     assert '3(5); 0:00:' in out
     assert '; CONVERGED' in out